Split devStore and prodStore
