+++ src/Utilities/getDevStore.js
@@ -1,0 +1,4 @@
+import { logger } from 'redux-logger';
+import { getStore } from './store';
+
+export const getDevStore = () => getStore([logger]);
