+++ src/entry-dev.js
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { DevEntry } from './entries';
+import { DevEntry } from './entries-dev';
 
 ReactDOM.render(<DevEntry />, document.getElementById('root'));
